feat(Error): Remove NewError() and add NewInvalidParamsErrorResponse(data interface{})
Users of the library rarely need to work with raw errors and responses
and really only need a convenience function for creating their own
InvalidParams error responses. This simplifies returning errors for
users MethodFuncs. It does, however change the function calls in a
non-backwards compatible way, so this will be part of a 2.0.0 release.

BREAKING CHANGE:
Removed:
- func NewError(code int, message string, data interface{}) Error
- func NewErrorResponse(err Error) Response
Added:
- func NewErrorResponse(code int, message string, data interface{}) Response
- func NewInvalidParamsErrorResponseError(data interface{}) Response