Update to v6 which improves registration of methods
Previously only a single JSON RPC server could be set up with a single
set of MethodFuncs using RegisterMethod(). Now the user provides a
MethodMap to a generator function which returns an http.Handler just for
that set of methods. This allows for reuse of this library to provide
multiple JSON RPC servers with distinct staticatlly defined sets of
methods.

Additionally this introduces changes that reduce unnecessary
unmarshalling of the ID and params and catch marshaling errors that
could be introduced by user data.

The http.Handler function has been refactored for clarity.

This also adds additional test coverage.

Finally this switches to lowercase file names.