fix(MethodFunc): Detect both Error and *Error, and panic on generic errors
Previously if a MethodFunc returned an Error or error it was being
returned to the client as the Response.Result instead of an Error. Now
both Error and *Error are detected and returned to the client as a Error
Response. Also if the MethodFunc returns a generic error, this should
not be considered as something to return to the client, and instead now
causes a panic and an InternalError to be returned to the client.