Don't fail immediately on bad HTTP status code
Try to parse the http.Response.Body as a Response regardless of
http.Response.StatusCode since some jsonrpc2 servers use status code AND
jsonrpc2.Error to indicate a method error. This is wrong but common.

Also, add ErrorUnexpectedHTTPResponse to allow deeper inspection of the
http.Response if parsing into a Response fails.

This slightly changes the error handling behavior for the
Client.Request().